swaybar: move core functionality to state.c
